+++ SDL/Triplet.cuh
@@ -26,11 +26,11 @@ namespace SDL
     struct triplets
     {
         unsigned int* segmentIndices;
-        unsigned int* lowerModuleIndices; //3 of them now
+        uint16_t* lowerModuleIndices; //3 of them now
         unsigned int* nTriplets;
 
         //for track extensions
-        unsigned int* logicalLayers;
+        uint8_t* logicalLayers;
         unsigned int* hitIndices;
         
         //delta beta = betaIn - betaOut
